Wacky Salmon config, and info of the last Push (Jun 22, 2022)
So, in this one, I added Config options for enabling the Salmon's random
size feature, and one for the wacky option for Salmon to grow in size
when fed.





For the previous push, I accidentally pushed without the full info, as I
keep doing Shift + Enter to try and goto a new line.

So, here is info on most everything added in tne PREVIOUS push. Ordered
by whatever file hierarchy is:


+Re-wrote all the Block and Item registration because most of it was
bad. Now it's easier to add to, is more centralized, can better handle
config stuff like disabling, and, again, it isn't as terrible.

+Part of this automatically links blocks to the Creative tab, so almost
every block had that stuff removed.

+Made it so Kelp can be grown up from any point of the stalk, rather
than the exact top block. Still might need to be tweaked in the future.

+Actually fixed Leaves to not decay if placed by the player!

+The top of Kelp can be interacted with using Shears to set the age to
max, so you can stop it from growing right where you want it!

+Removed useless bucket checks from the Cod and Salmon, bucketing has
been handled.

+You can use Bonemeal on Pickled to spawn Sea Pickles and... Hearts.
This stops their attacking, so it kinda stuns them. Also, if they are
names, the dropped Pickles inherit it with Jr attached.

+Drowned can literally be disabled via the config.

+Drowned properly rotate a held trident in hand!

+Drowned glowing 2ed layer can be disabled.

+Made the Drowned Armor Spawning, and exact spawn rates for held items
configuable! Now 100% of Drowned can have Tridents!

+Made Trident Drowned use melee if their target is within 3 blocks of
them. Range is configurable.

+Glow Squids light can be specifically tweaked in the config, including
their minBrightness, if they should have a 2ed layer, and the
minBrightness of the 2ed layer.

+Added BucketMobEvent, which handles the ability to use a bucket to
collect a mob. Accepts any mob in the config list.

+Squid Flipping can be disabled.

+Added PushSquidEvent, so the player can Shift Right-Click a Squid to
push it a bit. This is so you can acutally help a beached squid, and is
kinda funny. Can be disabled via config.

+The Trident can literally be disabled via the config. WHY, WHY NOT?!

+Heavily tweaked ItemSpawnBucket to work with BucketMobEvent.

+Changed the 'fake biome' generators to try and make them slightly more
seamless.

-Currently disabled Shore Rocks. They still consistently make cascading
worldgen lag, so options are being considered.

+Added a (admitedly intrusive) tweak to Water's light blocking.
Automatically disabled, use at your own risk.

+Sea Pickles should ONLY generate in 'Warm Ocean' areas

+Added Coral Reefs. Like, the whole things almost. Spawn within Warm
Oceans, always in the center.

+Added Coral Bulbs, the round ones exclusive to Java edition. The code
is a mess, I know.

+Added Coral Branches, the ones on their side reaching out.

+Added Coral Stalks, the ones sticking up. They actually generate with
branches at random points, like how Bedrock does!

+Changed up Coconut Tree generation to use the new wood, made the leaves
generate in a more interesting manner, and they have a base around the
bottom. It's just nice.

+Tweaked ALL worldgen to remove as much cascading worldgen as possible,
it is SO, SOO much better!

+Changed Coconut Saplings to Palm Saplings, because, ya know, it's a
palm tree.

+Added Palm Wood, which is generated by Coconut Trees!

+Added Palm WOOD, the non-log full-bark blocks.

+Added Palm Planks.

+Added Palm Stairs.

+Added Palm Slabs. They were annoying, so I used a brand new ItemBlock
for them.

+Added Palm Fences.

+Added Palm Fence Gates.

+Added Palm Doors.