Ocean Ruins & Locate Command
Added a temp structure with configuration options as well as a locate command to find it. Annotations have been left, and I am more than happy to work or get the Ocean Ruins where you want them. This push is just for you to see how it works and edit it how you like!